Its starting to look like an alarm clock
Added display methods for the clock and alarm. The serial still does
not work. Instead of using serial to set the clock. There will be a
button to do it. A menu system will be needed to include functions such
as set clock, set alarm, alarm on/off, and maybe even select
punishments.